+++ apps/vscode/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added diagnostics (i.e. squiggly underlines) to code cells in qmds (<https://github.com/quarto-dev/quarto/pull/980>)
 - Added setting and command to show/hide cells in outline (<https://github.com/quarto-dev/quarto/pull/974>).
 - Added custom pair colorization and highlighting for divs in qmds (<https://github.com/quarto-dev/quarto/pull/973>).
+- Added code symbols into outline (<https://github.com/quarto-dev/quarto/pull/972>).
 
 ## 1.132.0 (Release on 2026-05-05)